requirements.txt
asgiref==3.8.1
boto3==1.34.103
botocore==1.34.103
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-dotenv==1.4.2
django-filter==24.2
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jmespath==1.0.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
stripe==9.6.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
